Update specs for setOperator rule (#11)
This PR updates the specs that correspond to rule 2 about authorizing an account operator. While it only considered "the regular case" beforehand, it now also includes the case where msg.sender == currentContract (i.e. we are called via permit()) and the proper owner thus needs to be retrieved from the ExecutionContext.