Add inverse rule for setOperator (#14)
This PR adds a rule that @johspaeth and I wanted to add: the inverse for the existing setOperator rule, essentially saying "if the owner calls setOperator(), it will not revert". It turned out to be a bit more involved than expected, but discovered a few preconditions that might be worth adding to the EVC interface documentation.