Property 25+26: Transient storage is properly reset (#24)
This PR adds specs two more properties requested by Euler.
Property 25 states that the transient storage is properly reset after any top-level call. The transient property consists of the execution context, as well as the callback sets for account status checks and vault status checks.
Property 26 states that the execution context is restored to what it was before after any function call.
Instead of checking the various data in the execution context individually for property 25, we verify that the whole 256bit data is always fully reset to its default value, making the rule much cleaner.