Add reusable spec for call opcodes (#26)
We've been doing the same reasoning based on how delegatecall and callcode are used in several rules. This PR adds a reusable spec for these opcodes and explains the reasoning.
As we touch property 24 anyway, we do the same simplification as we do for the call and delegatecall hooks (direct assert instead of using a ghost) for the SetStorage hooks as well. This makes sure that this rule still works if ghosts get havoc'd in the future...